requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
python-doctr[torch]==0.6.0
torch==2.0.1
torchvision==0.15.2
Pillow==10.0.0
numpy==1.24.3
opencv-python==4.8.0.76
Werkzeug==2.3.7
gunicorn==21.2.0